main & usercontroller hotfix
