+++ src/main/java/com/example/recruit_page_wwy/HelloController.java
@@ -1,40 +1,17 @@
 package com.example.recruit_page_wwy;
 
 
-import com.example.recruit_page_wwy.user.User;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
-
-import com.example.recruit_page_wwy.employment.Employment;
 import com.example.recruit_page_wwy.employment.EmploymentRepository;
-import com.example.recruit_page_wwy.user.User;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
-
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
-
-import java.util.List;
 
 @RequiredArgsConstructor
 @Controller
 public class HelloController {
 
     private final EmploymentRepository employmentRepository;
-
-    @GetMapping("/")
-    public String index(HttpSession session, HttpServletRequest request) {
-        User sessionUser = (User) session.getAttribute("sessionUser");
-        request.setAttribute("sessionUser", sessionUser);
-        System.out.println(sessionUser);
-
-        List<Employment> jobs = employmentRepository.findTop4ByOrderByIdDesc();
-        request.setAttribute("models", jobs);
-
-        return "index";
-    }
 
     @GetMapping("/board/1")
     public String boardDetail() {
@@ -95,5 +72,5 @@ public class HelloController {
     public String applyManageList() {
         return "resume/com-apply-list";
     }
-    
+
 }
