+++ src/main/java/com/example/recruit_page_wwy/employment/EmploymentController.java
@@ -7,14 +7,29 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import java.util.List;
+
 @RequiredArgsConstructor
 @Controller
 public class EmploymentController {
     private final EmploymentService employmentService;
     private final HttpSession session;
 
+    @GetMapping("/")
+    public String index(HttpSession session, HttpServletRequest request) {
+        User sessionUser = (User) session.getAttribute("sessionUser");
+//        request.setAttribute("sessionUser", sessionUser);
+//        System.out.println(sessionUser);
+
+        List<Employment> jobs = employmentService.viewEmployList();
+        request.setAttribute("models", jobs);
+
+        return "index";
+    }
+
     @GetMapping("/mypage/employment")
     public String manageEmployment(HttpServletRequest request) {
+
 
         // 마이페이지에서 아직 세션 정보 불러와지지 않음. 유저id 임시로 4 줌
         //User sessionUser = (User) session.getAttribute("sessionUser");
