+++ src/main/java/com/example/recruit_page_wwy/employment/EmploymentService.java
@@ -23,6 +23,7 @@ public class EmploymentService {
 
         return dtoList;
     }
+
     public List<EmploymentResponse.PublicListDTO> emplymentAllList(Integer userId) {
         List<Employment> employmentAllList = employmentRepository.findAll();
 
@@ -33,5 +34,9 @@ public class EmploymentService {
         }
 
         return dtoList;
+    }
+
+    public List<Employment> viewEmployList() {
+        return employmentRepository.findTop4ByOrderByIdDesc();
     }
 }
