+++ src/main/java/com/example/recruit_page_wwy/user/UserController.java
@@ -88,3 +88,4 @@ public class UserController {
         session.invalidate();
         return "redirect:/";
     }
+}
